pyproject.toml
# SPDX-License-Identifier: Apache-2.0
# Copyright 2026 Mark Eatherly

[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[project]
name = "qwashed"
version = "0.1.0"
description = "Free post-quantum hygiene for civil society: HNDL audit + hybrid-PQ vault."
readme = "README.md"
license = { text = "Apache-2.0" }
authors = [
    { name = "Mark Eatherly" },
]
maintainers = [
    { name = "Mark Eatherly" },
]
requires-python = ">=3.11"
keywords = [
    "post-quantum",
    "cryptography",
    "civil-society",
    "hndl",
    "pqc",
    "audit",
    "vault",
    "ml-kem",
    "ml-dsa",
    "hybrid",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "Intended Audience :: Information Technology",
    "Intended Audience :: Legal Industry",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: MacOS",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security :: Cryptography",
    "Topic :: System :: Networking :: Monitoring",
    "Typing :: Typed",
]

# Core dependencies (always installed). The audit and vault modules each
# have their own optional-dependency extras so the base install stays small
# and air-gappable for users who only need the core APIs.
dependencies = [
    "pydantic>=2.5,<3.0",
    "cryptography>=42.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
# `qwashed audit` (HNDL Auditor) extras
audit = [
    "sslyze>=6.0",
    "paramiko>=3.4",
]

# `qwashed vault` (Hybrid PQ Vault) extras
# liboqs-python wraps the upstream liboqs C library (NIST FIPS 203/204).
# On macOS arm64 / Linux x86_64 a wheel is published; otherwise liboqs
# must be available on the system (Homebrew: `brew install liboqs`).
vault = [
    "liboqs-python>=0.14.0",
    "argon2-cffi>=23.1",
]

# Optional PDF report generation
report = [
    "reportlab>=4.0",
]

# Convenience: install everything user-facing
full = [
    "qwashed[audit,vault,report]",
]

# Development and CI
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "hypothesis>=6.100",
    "ruff>=0.5",
    "mypy>=1.10",
    "types-pyyaml>=6.0",
    "types-paramiko>=3.4",
]

[project.scripts]
qwashed = "qwashed.cli:main"

[project.urls]
Homepage = "https://github.com/qwashed/qwashed"
Documentation = "https://github.com/qwashed/qwashed/blob/main/docs/QUICKSTART.md"
Repository = "https://github.com/qwashed/qwashed"
Issues = "https://github.com/qwashed/qwashed/issues"
"Threat Model" = "https://github.com/qwashed/qwashed/blob/main/THREAT_MODEL.md"
"Build Plan" = "https://github.com/qwashed/qwashed/blob/main/QWASHED_BUILD_PLAN.txt"

# ---------------------------------------------------------------------------
# Hatch build configuration
# ---------------------------------------------------------------------------

[tool.hatch.build.targets.wheel]
packages = ["qwashed"]

[tool.hatch.build.targets.sdist]
include = [
    "/qwashed",
    "/tests",
    "/docs",
    "/examples",
    "/LICENSE",
    "/NOTICE",
    "/README.md",
    "/CHANGELOG.md",
    "/THREAT_MODEL.md",
    "/QWASHED_BUILD_PLAN.txt",
    "/pyproject.toml",
]

# ---------------------------------------------------------------------------
# Ruff (lint + format)
# ---------------------------------------------------------------------------

[tool.ruff]
line-length = 100
target-version = "py311"
extend-exclude = ["docs/_build", ".venv", "build", "dist"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "W",    # pycodestyle warnings
    "B",    # flake8-bugbear
    "UP",   # pyupgrade
    "S",    # flake8-bandit (security)
    "RUF",  # ruff-specific
    "SIM",  # flake8-simplify
    "C4",   # flake8-comprehensions
    "PTH",  # flake8-use-pathlib
]
ignore = [
    # We use assert in tests; unit tests are not security-critical paths.
    # (Per-file ignore below covers tests/, but keep this off in src too
    # because Pydantic validators sometimes assert invariants.)
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "S101",  # asserts allowed in tests
    "S105",  # hardcoded passwords allowed in tests (fixtures)
    "S106",  # hardcoded passwords allowed in tests
    "B017",  # pydantic ValidationError is subclass of Exception; broad raise OK in tests
    "SIM105",  # nested try/except/pass blocks in fixture teardown are clearer than contextlib.suppress
    "S603",  # controlled subprocess.run([sys.executable, ...]) in smoke tests is safe
]
"qwashed/cli.py" = [
    "S603",  # subprocess call allowed in CLI helpers (none in v0.1)
]
"qwashed/audit/cli.py" = [
    "S101",  # parse_strict invariant narrower (mypy)
]
"qwashed/audit/profile_loader.py" = [
    "S101",  # parse_strict invariant narrower (mypy)
]
"qwashed/audit/probe.py" = [
    "S110",   # best-effort sslyze metadata enrichment is intentionally silent
    "SIM117", # nested with-blocks make the TLS handshake flow easier to read
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# ---------------------------------------------------------------------------
# mypy (static type checking)
# ---------------------------------------------------------------------------

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_ignores = true
warn_return_any = true
warn_redundant_casts = true
warn_unreachable = true
no_implicit_optional = true
disallow_untyped_defs = true
disallow_any_generics = true
check_untyped_defs = true

[[tool.mypy.overrides]]
module = [
    "sslyze.*",
    "paramiko.*",
    "oqs.*",
    "argon2.*",
    "reportlab.*",
]
ignore_missing_imports = true

# ---------------------------------------------------------------------------
# pytest
# ---------------------------------------------------------------------------

[tool.pytest.ini_options]
minversion = "8.0"
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--showlocals",
]
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "network: marks tests that require network access",
    "kat: marks NIST FIPS 203/204 known-answer-test vectors",
]

[tool.coverage.run]
source = ["qwashed"]
branch = true
omit = [
    "qwashed/__main__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
